# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma de Subastas en Línea — API
  version: "1.0.0-entrega1"
  description: >
    API REST de la Entrega 1. Cubre la gestión de usuarios, la publicación y
    consulta de subastas y el registro de pujas. La comunicación en tiempo real
    y el webhook de pagos corresponden a entregas posteriores.

    Los valores monetarios son enteros no negativos en pesos colombianos. Las
    reglas de negocio se validan en el dominio: una puja rechazada por una regla
    de negocio nunca responde con un código de éxito.
servers:
  - url: http://localhost:3000
    description: Entorno local
...
paths:
...
  /subastas:
    get:
      tags: [Subastas]
      summary: Listar subastas con filtros y paginación
      security: []
      parameters:
        - in: query
          name: categoriaId
          schema:
            type: string
        - in: query
          name: estado
          schema:
            $ref: "#/components/schemas/EstadoSubasta"
        - in: query
          name: pagina
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: tamanoPagina
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        "200":
          description: Página de subastas
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  subastas:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subasta"
    post:
      tags: [Subastas]
      summary: Publicar una subasta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [denominacion, descripcion, estadoConservacion, categoriaId, precioBase, incrementoMinimo, fechaCierre]
              properties:
                denominacion:
                  type: string
                  example: Reloj de bolsillo antiguo
                descripcion:
                  type: string
                  example: Pieza de colección en buen estado de conservación
                estadoConservacion:
                  $ref: "#/components/schemas/EstadoConservacion"
                categoriaId:
                  type: string
                precioBase:
                  type: integer
                  minimum: 1
                  example: 100000
                incrementoMinimo:
                  type: integer
                  minimum: 1
                  example: 10000
                fechaCierre:
                  type: string
                  format: date-time
                  description: >
                    Posterior al momento de publicación; la duración debe estar
                    entre una hora y treinta días.
      responses:
        "201":
          description: Subasta publicada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subasta"
        "400":
          description: >
            Datos con formato inválido, o alguna condición de publicación no
            se cumple (precio base o incremento no positivo, fecha de cierre
            anterior a la publicación, duración fuera de rango).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "404":
          description: La categoría indicada no existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    tokenPortador:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NoAutenticado:
      description: Falta el token de sesión o es inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            codigo:
              type: string
              description: Código identificable del error
              example: RN09_PUJA_INSUFICIENTE
            mensaje:
              type: string
              example: La puja debe superar la oferta vigente en al menos el incremento mínimo.
            detalles:
              type: object
              additionalProperties: true
              description: Presente solo en errores de validación de formato
...
    EstadoConservacion:
      type: string
      enum: [NUEVO, COMO_NUEVO, BUEN_ESTADO, ACEPTABLE, PARA_REPARAR]
    EstadoSubasta:
      type: string
      enum: [ABIERTA, CANCELADA, ADJUDICADA, DESIERTA]
    Articulo:
      type: object
      properties:
        denominacion:
          type: string
        descripcion:
          type: string
        estadoConservacion:
          $ref: "#/components/schemas/EstadoConservacion"
        categoriaId:
          type: string
...
    OrdenPago:
      type: object
      nullable: true
      properties:
        id:
          type: string
        ganadorId:
          type: string
        monto:
          type: integer
        fechaLimite:
          type: string
          format: date-time
        estado:
          type: string
          enum: [PENDIENTE, CONFIRMADA, VENCIDA]
    Subasta:
      type: object
      properties:
        id:
          type: string
        vendedorId:
          type: string
        articulo:
          $ref: "#/components/schemas/Articulo"
        precioBase:
          type: integer
        incrementoMinimo:
          type: integer
        fechaPublicacion:
          type: string
          format: date-time
        fechaCierre:
          type: string
          format: date-time
        estado:
          $ref: "#/components/schemas/EstadoSubasta"
        ofertaVigente:
          type: integer
          nullable: true
        mejorPostorId:
          type: string
          nullable: true
        cantidadPujas:
          type: integer
        ordenPago:
          $ref: "#/components/schemas/OrdenPago"
security:
  - tokenPortador: []